requirements.txt
python-escpos
pyusb
libusb